requirements.txt
streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
openai>=1.30.0